chore(FakeProfanityChecker): Move directory
